Adding a license.
