+++ phantom-connectors/src/main/scala/com/outworkers/phantom/connectors/package.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2015 Sphonic Ltd. All Rights Reserved.
+ * Copyright 2013 - 2017 Outworkers Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
