--- phantom-example/src/test/scala/com/websudos/phantom/example/ExampleSuite.scala
@@ -1,68 +1,0 @@
-/*
- * Copyright 2013-2016 Websudos, Limited.
- *
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * - Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * - Redistributions in binary form must reproduce the above copyright
- * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution.
- *
- * - Explicit consent must be obtained from the copyright owner, Outworkers Limited before any redistribution is made.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-package com.outworkers.phantom.example
-
-import com.outworkers.phantom.connectors.RootConnector
-import com.outworkers.phantom.Manager._
-import com.outworkers.phantom.example.advanced.RecipesDatabase
-import com.outworkers.util.lift.{DateTimeSerializer, UUIDSerializer}
-import org.scalatest.concurrent.{PatienceConfiguration, ScalaFutures}
-import org.scalatest._
-
-import scala.concurrent.Await
-import scala.concurrent.duration._
-
-trait BaseSuite extends Suite
-  with Matchers
-  with BeforeAndAfterAll
-  with RootConnector
-  with ScalaFutures
-  with OptionValues {
-
-  implicit val defaultTimeout: PatienceConfiguration.Timeout = timeout(10 seconds)
-
-  implicit val formats = net.liftweb.json.DefaultFormats + new UUIDSerializer + new DateTimeSerializer
-}
-
-
-trait ExampleSuite extends Suite with BaseSuite with RecipesDatabase.connector.Connector {
-
-  override def beforeAll(): Unit = {
-    super.beforeAll()
-    Await.ready(RecipesDatabase.autocreate().future(), 20.seconds)
-  }
-
-  override def afterAll(): Unit = {
-    super.afterAll()
-    Await.result(RecipesDatabase.autotruncate().future(), 20.seconds)
-  }
-}
-
-
